Flatten directory
